requirements.txt
Django==5.0.2
djangorestframework==3.14.0
requests==2.32.5
Pillow==10.0.0
python-decouple==3.8
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
